+++ src/dspf-edit.commands/dspf-edit.add-attribute.ts
@@ -8,7 +8,7 @@ import * as vscode from 'vscode';
 import { DdsNode } from './../dspf-edit.providers/dspf-edit.providers';
 import { isAttributeLine, findElementInsertionPoint } from './../dspf-edit.utils/dspf-edit.helper'
 import { ExtensionState } from './../dspf-edit.states/state';
-import { fieldsPerRecords } from './../dspf-edit.parser/dspf-edit.model';
+import { fieldsPerRecords } from '../dspf-edit.model/dspf-edit.model';
  
 // INTERFACES AND TYPES