+++ src/dspf-edit.commands/dspf-edit.add-buttons.ts
@@ -6,7 +6,7 @@
 
 import * as vscode from 'vscode';
 import { DdsNode } from './../dspf-edit.providers/dspf-edit.providers';
-import { getRecordSize, fieldsPerRecords, DdsSize, getDefaultSize } from './../dspf-edit.parser/dspf-edit.model';
+import { getRecordSize, fieldsPerRecords, DdsSize, getDefaultSize } from '../dspf-edit.model/dspf-edit.model';
 import { ExtensionState } from './../dspf-edit.states/state';
 
 // COMMAND REGISTRATION