+++ CHANGELOG.md
@@ -99,3 +99,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.6.1] - 2025-09-09
 ## Fixes
 - Internal fixes. Refactoring.
+- The resizing function was not working correctly.